7/10

+++ database/migrations/2023_09_30_075930_create_order_payment_methods_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
 
             $table->string('payment_provider', 255)->nullable();
             $table->string('status', 255)->nullable();
-            $table->float('total')->nullable()->unsigned();
+            $table->double('total')->nullable()->unsigned();
             $table->string('note', 255)->nullable();
 
             $table->softDeletes();
